fix `container_exec_cmd` calls for --limit runs
When ceph-ansible is run with `--limit`, a call to
`hostvars[groups[mon_group_name][0]]['container_exec_cmd']` will fail if
there's no monitor present in the current play because `ceph-facts`
hasn't been run on a mon.

Forcing `container_exec_cmd` to be set on each mon when it's needed
fixes this issue.

Closes: https://bugzilla.redhat.com/show_bug.cgi?id=1792320

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>